### 2019-12-20 #1973 @whedon
[PRE REVIEW]: ReferenceSeeker: rapid determination of appropriate reference genomes
**Submitting author:** @oschwengers (<a href="http://orcid.org/0000-0003-4216-2721">Oliver Schwengers</a>)
**Repository:** <a href="https://github.com/oschwengers/referenceseeker" target ="_blank">https://github.com/oschwengers/referenceseeker</a>
**Version:** v1.4
