### 2019-12-20 #1973 @labarba
Please note that `@mgymrek` has stepped down from the editorial board. I'm the associated EiC on rotation this week, and am looking for another handling editor.
