### 2019-12-20 #1973 @labarba
👋 @lpantano — Could you handle this submission as editor?
Have look please. We understand of course that these days will be slow and you might be on holiday. Let us know, and we can assign you for handling this after the New Year.
