### 2019-12-21 #1973 @lpantano
it seems super interesting but I am ok if @will-rowe can take it bc i’ll be away for two weeks! i will follow the thread!
