### 2019-12-21 #1973 @labarba
Thanks for the offer, @will-rowe — I'll assign this to you, and look for someone else for https://github.com/openjournals/joss-reviews/issues/1972
