### 2020-01-06 #1973 @standage
I'd be happy to. Note: I have another JOSS review in progress, so it may be a couple weeks before I can get to this.
