### 2020-01-20 #1994 @oschwengers
@will-rowe Thanks a lot for accepting the [ReferenceSeeker ](https://github.com/oschwengers/referenceseeker) editor role.

@luizirber @standage Thank you very much for agreeing to review my work and for the first checkmarks.

If there's anything I can do that may further help you in the reviewing process, please let me know.
