### 2020-01-28 6021f3f
add support for bidirectional ANI/conDNA values
