### 2020-01-28 #1994 @oschwengers
I briefly mentioned the new support for `bidirectional` computations of `ANI` and `conserved DNA` values in the manuscript.
