### 2020-01-29 oschwengers/referenceseeker#4 @standage
Test suite or instructions
Great package, thanks for making this available!

Surely you have tested ReferenceSeeker extensively to make sure it performs as expected on numerous inputs. However, unless I'm mistaken, the documentation doesn't include any instructions for running an automated test suite or commands to execute to ensure the software is installed correctly.

Bundling entire bacterial/viral/fungal/etc. components of RefSeq with the software to support an automated test suite isn't feasible—it would require an incredible amount of storage for an otherwise lightweight package. However, one possible approach would be to construct a smaller mock database with a small number of entries that could be more easily bundled with ReferenceSeeker.

At the very least, beginners would benefit from 1 or 2 test examples that very clearly demonstrate:

- how to download the input data
- how to run the test
- what the expected output should be
