### 2020-01-29 oschwengers/referenceseeker#5 @standage
Continuous integration
Following up on #4, if ReferenceSeeker had an automated test suite, it would also be possible to go a step further and set up a continuous integration environment using a service like GitHub Actions.

- While software is under active development, CI services can be configured to run a test suite and time updates are made to the code (such as from a push to `master` or a pull request from the community). The repo can also be configured so that any pull request submitted by the core developer(s) or the community cannot be merged until the tests pass.
- If/when software is in maintenance, CI services can be configured to run a test suite on a schedule, such as once per week or once per month. If future updates to package dependencies break ReferenceSeeker, CI would ensure a notification is issued to the authors early so that the problem can be addressed quickly. This would make it possible for the authors to provide long-term support for the software with minimal effort.
