### 2020-01-29 oschwengers/referenceseeker#6 @standage
Feature request: ability to include private data
Consider two hypothetical scenarios.

- a food-borne illness lab with numerous well-annotated *Salmonella* genomes beyond what's available in RefSeq
- an infectious disease lab with numerous well-annotated *Yersinia pestis* strains not publicly available

It would be helpful to scientists in labs like these to be able to include private data along with RefSeq genomes, either during the build process or by adding them later.
