### 2020-01-29 #1994 @standage
The authors describe ReferenceSeeker, a Python package designed to select the most appropriate reference genome given a draft genome assembly. ReferenceSeeker implements a two-stage selection procedure: in the first, candidate reference genomes (CRGs) are identified using rapid k-mer based operations; in the second, it performs more computationally intensive alignment-based operations on a small number of CRGs to identify the reference genome (RG) most closely related to the input. The harmonic mean of average nucleotide identity (ANI) and conserved DNA (conDNA) statistics is the final selection and ranking criterion. These two statistics strike a balance between sequence similarity and reciprocal coverage as a measure of distance.

ReferenceSeeker addresses a fundamental preliminary step common to many comparative genomics studies, the results of which have a substantial impact on interpretation of downstream results. The tool furnished by the authors is well motivated, well described, and well implemented.

- The software package is straightforward to install. I was able to install it successfully on an airgapped HPC system without using root privileges, with a reasonable amount of effort. The typical user will have an even easier time installing ReferenceSeeker with the suggested `conda install` method.
- ReferenceSeeker requires pre-computed databases. The authors provide hyperlinks to stable archived snapshots of these required databases, but (critically) also provide the scripts necessary to build those scripts from scratch in the future.
- The software package has an easy-to-use and self-documenting interface. Invoking `referenceseeker --help` lists the expected inputs and how to adjust similarity thresholds and configure other aspects of the tool's behavior.
- In my hands, the tool performs as advertised. I randomly selected five private samples (microbial pathogens) for testing.
    - In three cases, the top ReferenceSeeker result identified the correct species and strain reference.
    - In one case, the top ReferenceSeeker results identified the correct species reference. The actual strain was unknown, but the ReferenceSeeker results appear quite reasonable compared to alternative approaches we've previously used for classification.
    - In one case, the top ReferenceSeeker result identified the correct genus but the species call was incorrect. To be fair, this was a *Shigella* genome, which we have had a notoriously difficult time classifying correctly. The correct strain was within the top 10 results, which included various *Shigella* and *E. coli* strains. This is comparable to the performance of many leading tools on this particularly problematic sample.

This tool is a valuable contribution to the field, and I am enthusiastic to continue using it myself. However, I have two concerns that should be addressed before I can endorse ReferenceSeeker for publication at JOSS.

## Major concern: precise definitions of statistics used

The terms ANI and "conserved DNA" are used throughout the paper and technical documentation, but no precise definitions are provided. The paper containing definitions for these statistics is cited throughout the manuscript, but never as a reference for defining conserved DNA.

- The first reference appears in the **Summary** paragraph, but refers only to ANI.
- The second reference appears in the **Determination of RG** section, but refers to species boundaries, and is written as if "conDNA" has already been defined.

Many readers will be familiar with ANI and conserved DNA, but many will not, and multiple formulations of these statistics exist. The authors need to make it unambiguously clear how these statistics are computed by ReferenceSeeker. Please ensure that the connection between conDNA and (Goris *et al.*, 2007) is clear. I would even go so far as to suggest that explicit descriptions of how the statistics are calculated should be included, presumably in the **Determination of RG** section. Something like the following might work.

> ReferenceSeeker uses both ANI and percentage of conserved DNA (conDNA) to derive a highly specific measure of microbial genome relationships (Goris *et al.*, 2007).
> ANI is calculated as the mean identity of all nucmer matches that <*...fill in remaining details...*>.
> The conDNA statistic is calculated as the sum of <*...fill in remaining details...*>.
> The Mash distances used for preliminary selection of CRGs correlate well with ANI—these statistics capture nucleotide-level *sequence similarity*.
> However, Mash distances do not correlate well with the conDNA statistic, which captures *reciprocal coverage* of the query and reference sequences.
> The final RG selection criterion incorporates both ANI and conDNA to ensure that both sequence similarity and reciprocal coverage are considered when ranking RGs.

This may require editing/revising of the entire section.

## Major concern: no automated test suite

Unless I missed something, there doesn't appear to be any automated test suite, or even a description of manual steps for testing the software. Tests serve at least two important purposes.

1. They allow users to independently verify that the software has been installed correctly and is running as expected on their system.
2. They allow developers to evaluate whether any proposed changes to the code (bug fixes, new features, refactoring, removing vestigial code, etc.) will have any unintended consequences.

There can be no doubt that the authors have been responsible in testing ReferenceSeeker: it would not function effectively if they had not. But it seems appropriate for publication in this venue, and as common practice, to encode some of these tests in an automated test suite. At the **very minimum**, users should have instructions they can manually run to test the software, but an automated test suite is *strongly encouraged*. (The [pytest](https://docs.pytest.org/en/latest/) framework is the *de facto* standard for testing in the Python community.)

With an automated test suite, it would also be possible to go a step further and set up a continuous integration environment using a service like GitHub Actions, Travis CI, Drone.io, or CircleCI. CI services can be configured to run a test suite on a schedule (such as once per week or once per month) and/or when any updates are made to the code (such as from a push to `master` or a pull request from the community). This would alert the authors any time, for example, an update to a package dependency breaks ReferenceSeeker's behavior, helping to ensure ReferenceSeeker remains usable for more than just a short time following publication. **Note**: this should not be taken as a strict requirement for acceptance, but a stretch goal for increasing ReferenceSeeker's value over time.

Related threads:

- https://github.com/oschwengers/referenceseeker/issues/4
- https://github.com/oschwengers/referenceseeker/issues/5

## Suggestion: ability for users to include private data

In addition to these major concerns, I also have a minor concern regarding the ability to include private data sets. Imagine for example a food-borne illness lab with numerous well-annotated *Salmonella* genomes beyond what's available in RefSeq, or an infectious disease lab with numerous well-annotated *Yersinia pestis* strains not publicly available. It would be valuable to scientists working in such labs to be able to include their private data along with the RefSeq genomes during the database build process, or to add them later. **Note**: this should also not be taken as a requirement for acceptance, but a suggestion for added value.

Related threads:

- https://github.com/oschwengers/referenceseeker/issues/6
