### 2020-01-29 #1994 @oschwengers
@standage Thank you so much for this excitingly comprehensive & constructive feedback!

- I adopted the `ANI` & `conDNA` calculation from `Goris et al., 2007` but I totally agree that an exact explanation and formula is more than appropriate, especially since the formula is somehow "hidden" in prosa in the original article.
- Also I agree that at least a tiny test for the users or even an automated one (with CI) would be a true benefit.
- Just recently, I discussed the support for private data with @nick60 but postponed it in first place. So I guess I'll just bring it forward.

I'm more than happy to implement all comments and suggestions.
