### 2020-01-29 def7682
remove unnecessary itertools dependency
