### 2020-01-30 f3dc942
add check for empty query files
