### 2020-01-30 74a24fb
polish cli usage and readme
