### 2020-01-30 84b1458
add Salmonella spp mock db and test case #4
