### 2020-01-30 858267d
Add first PyTest unit tests #5
