### 2020-01-30 3240043
Add CI python tests on pushs #5
