### 2020-01-30 3d24d80
Include deps in ci installation #5
