+++ .github/workflows/pythonpackage.yml
@@ -22,7 +22,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install --no-deps --ignore-installed .
+        pip install --ignore-installed .
     - name: Lint with flake8
       run: |
         pip install flake8
