### 2020-01-30 946d3fa
Expand CI to pull requests #5
