+++ .github/workflows/pythonpackage.yml
@@ -1,6 +1,6 @@
 name: Python package
 
-on: push
+on: [push, pull_request]
 
 jobs:
   build:
