### 2020-01-30 3e0669f
Expand CI to different OS
