+++ .github/workflows/pythonpackage.yml
@@ -5,10 +5,10 @@ on: [push, pull_request]
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-16.04, ubuntu-18.04, ubuntu-latest, macos-latest]
         python-version: [3.5, 3.6, 3.7, 3.8]
 
     steps:
