### 2020-01-30 5fef44f
rewrite 'determination of RG' section
