### 2020-01-30 oschwengers/referenceseeker#4 @oschwengers
Hi @standage , thanks a lot for this!
You're absolutely right, so far there hasn't been any test functionality mainly due to the large databases as you've already pinpointed it.

I therefore created a tiny mock db comprising only 4 *Salmonella spp* genomes as well as a test query genome from SRA and added this to a new `tests` directory.

Accordingly, I added a short `Test` subsection to the `Installation` section of the readme.

As for now, I'd rather stick to this simple test case not to flood the readme too much.
Would you consider this as sufficient and comprehensible or would you like to see even more test cases?
