### 2020-01-30 oschwengers/referenceseeker#5 @oschwengers
Hi @standage , thanks for your comment!

I added a new `tests` directory comprising 2 new test modules comprising 9 unit-tests testing:
- the calculation of `ANI` and `conDNA` values
- the creation of to-be-aligned nucleotide fragments
to start with.

Right now, I cannot come up with more unit tests which could be added without having to refactor major parts of the software. Any ideas/comments very much appreciated!

Additionally, I setup a Python package CI workflow via GitHub Actions which triggers:
- packaging
- testing
the software on all combinations of:
- OS: ubuntu-16.04, ubuntu-18.04, ubuntu-latest, macos-latest
- Python versions: 3.5, 3.6, 3.7, 3.8

Just in case I missed something, please, don't hesitate to ask. If not, I'd close this issue.
