### 2020-01-30 oschwengers/referenceseeker#5 @oschwengers
@standage Once again for pointing to GitHub actions! I just added an `PyPI` deploy action - might save some time on future releases.
