### 2020-01-30 #1994 @standage
Hi @will-rowe, my main concerns have been addressed, and I endorse this paper for publication at JOSS! This concludes my review.

Hi @oschwengers, thanks for sharing your work. Feel free to reach out in the future if needed.
