+++ referenceseeker/database.py
@@ -1,0 +1,183 @@
+
+import argparse
+from pathlib import Path
+import shutil
+import subprocess as sp
+import sys
+import tempfile
+
+from Bio import SeqIO
+
+import referenceseeker
+import referenceseeker.constants as rc
+import referenceseeker.util as util
+
+
+def init(args):
+    try:
+        cwd_path = Path(args.output).resolve()
+        db_path = cwd_path.joinpath(args.db)
+        if(db_path.exists()):
+            print("detected existing database directory: %s" % db_path)
+            if(not db_path.is_dir()):
+                sys.exit("Database path (%s) exists but is not a directory!" % db_path)
+        else:
+            db_path.mkdir(mode=0o770)
+            print("created database directory")
+
+        db_tsv_path = db_path.joinpath('db.tsv')
+        db_tsv_path.touch(mode=0o660)
+        print("created database genome information file (db.tsv)")
+
+        db_sketch_path = db_path.joinpath('db.msh')
+        db_sketch_path.touch(mode=0o660)
+        print("created database genome kmer sketch file (db.msh)")
+    except:
+        print("Error: could not init database (%s) in output directory (%s)!" % (args.output, args.db), file=sys.stderr)
+        raise
+        sys.exit(-1)
+    print("\nSuccessfully initialized empty database at %s" % db_path)
+    print("Use 'referenceseeker_db import' to import genomes into database")
+
+
+def import_genome(args):
+    try:
+        db_path = Path(args.db).resolve()
+        genome_path = Path(args.genome).resolve()
+        tmp_path = Path(tempfile.mkdtemp())
+        genome_suffix = genome_path.suffix.lower()
+        genome_id = args.id
+        if(genome_suffix in ['.fasta', '.fas', '.fsa', '.fna', 'fa']):
+            # import fasta
+            pass
+        elif(genome_suffix in ['.genbank', '.gbff', 'gbk', 'gb']):
+            # import genbank
+            input_path = genome_path
+            genome_path = tmp_path.joinpath('genome.fasta')
+            with input_path.open() as fh_in, genome_path.open('w') as fh_out:
+                sequences = SeqIO.parse(fh_in, "genbank")
+                SeqIO.write(sequences, fh_out, "fasta")
+        elif(genome_suffix in ['.embl', '.ebl', 'el']):
+            # import embl
+            input_path = genome_path
+            genome_path = tmp_path.joinpath('genome.fasta')
+            with input_path.open() as fh_in, genome_path.open('w') as fh_out:
+                sequences = SeqIO.parse(fh_in, "embl")
+                SeqIO.write(sequences, fh_out, "fasta")
+        else:
+            raise Exception("Unknown genome file extension (%s)" % genome_suffix)
+
+        # extract genome id if it is empty
+        if(genome_id is None):
+            for record in SeqIO.parse(str(genome_path), 'fasta'):
+                genome_id = record.id
+                break
+
+        # copy genome fasta file to database directory
+        shutil.copyfile(str(genome_path), str(db_path.joinpath("%s.fna" % genome_id)))
+
+        # sketch genome
+        cmd = [
+            'mash',
+            'sketch',
+            '-o', 'genome',
+            '-k', '32',
+            '-s', '10000',
+            str(genome_path)
+        ]
+        proc = sp.run(
+            cmd,
+            cwd=str(tmp_path),
+            stdout=sp.PIPE,
+            stderr=sp.PIPE,
+            universal_newlines=True
+        )
+        if(proc.returncode != 0):
+            sys.exit("ERROR: failed to create genome kmer sketches via Mash!\nexit=%d\ncmd=%s" % (proc.returncode, cmd))
+
+        existing_db_sketch_path = db_path.joinpath('db.msh')
+        if(existing_db_sketch_path.stat().st_size == 0):
[…]
