+++ setup.py
@@ -30,7 +30,8 @@ setup(
     ],
     entry_points={
         'console_scripts': [
-            'referenceseeker=referenceseeker.referenceseeker:main'
+            'referenceseeker=referenceseeker.referenceseeker:main',
+            'referenceseeker_db=referenceseeker.database:main'
         ]
     },
     classifiers=[
