### 2020-01-30 oschwengers/referenceseeker#6 @oschwengers
Hi @standage , @nick60 ,
I just implemented a new database `init` and `import` logic and added a new `referenceseeker_db` cli. A brief description was also added to the database readme section.

I conducted some preliminary but successful tests so feel free to give it try.

As for now, it's only on GitHub but I'll put it on PyPI & BioConda as soon as I did some more tests.
