### 2020-01-31 #1994 @oschwengers
Hi @standage , once again thank you very much for your very constructive and helpful review. I adopted the manuscript as requested and implemented all other suggestions:
- a tiny test case to check local software installations: oschwengers/referenceseeker/issues/4
- added unit tests and setup CI via GitHub actions: oschwengers/referenceseeker/issues/5
- custom database creations: oschwengers/referenceseeker/issues/6

A last check mark on `software paper` - `summary` is left over - is there still anything missing you would like to see revised or added?
