### 2020-01-31 e3e7f1e
fix import genome file suffixes
