### 2020-01-31 oschwengers/referenceseeker#6 @oschwengers
@standage There was a tiny but nasty bug lurking in the database creation code. I did some further tests today and now, everything seems to work.

I'll close this issue and create a new release containing all recent changes. If you'll be facing any bugs or issues related to this, please feel free to reopen this issue or open a new one.
