### 2020-01-31 4cdf513
Fix PyPI upload action
removed If condition and added user: __token__
