### 2020-01-31 #1994 @oschwengers
@will-rowe  I just released [v1.5](https://github.com/oschwengers/referenceseeker/releases/tag/v1.5) containing all recent changes & updates related to @standage 's fruitful comments. v1.5 is already deployed to `PyPI` and `BioConda`. Is it possible to point this submission/review to this new release instead of v1.4?
