### 2020-01-31 32d6ee1
fix 2nd typo in manuscript
