### 2020-01-31 #1994 @oschwengers
Just fixed 2 typos.
