### 2020-01-31 cfac1ab
add note to manuscript on new database import features
