### 2020-02-03 #1994 @will-rowe
Hi @oschwengers - thanks for addressing the reviewer comments so quickly.

I've just checked the docs and can't find any way to change the version during review. Also, @luizirber may have already done his review for the previous version? Once we have @luizirber's review and his comments are addressed, we will then ask you to tag a [new release](https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance). I think the best we can do for now is to make Luiz aware that you have already addressed Daniel's comments.
