### 2020-02-03 #1994 @luizirber
This looks great, and @standage already did a stellar job in his review, so just some things I missed:

## Functionality

- Installation: Test section has a typo in commands (`test` instead of `tests`), this works:
  * `referenceseeker referenceseeker/tests/db referenceseeker/tests/Salmonella_enterica_CFSAN000189.fasta`
  * `./referenceseeker/bin/referenceseeker referenceseeker/tests/db referenceseeker/tests/Salmonella_enterica_CFSAN000189.fasta`

## Documentation

- Community guidelines: I didn't find instructions for contributing. Maybe add a [CONTRIBUTING file](https://blog.github.com/2012-09-17-contributing-guidelines/), and instructions to run tests (I dug into the GitHub Actions to be sure it was `pytest` that is running them).
  A stretch goal would be to add [issue and PR templates](https://github.blog/2016-02-17-issue-and-pull-request-templates/),  requesting some info that is useful for you to reproduce problems (for issues) and a checklist for new contributions (in PRs).
