+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,18 @@
+---
+name: Bug report
+about: Create a report to help us improving the ReferenceSeeker
+title: "[BUG]"
+labels: bug
+assignees: ''
+
+---
+
+**Describe the bug**
+To help users and fix any bugs and issues a concise description if beneficial and often necessary.
+
+Therefore, please provide us with at least the following information:
+- what exactly happened
+- what exact command was executed: just copy-paste the command line
+- what installation of ReferenceSeeker did you use: `BioConda`, `GitHub`, `Pip`
+- what database was used: `public`, `custom`, which taxon
+- which version of ReferenceSeeker was used
