### 2020-02-04 #1994 @oschwengers
Hi @luizirber  , thanks a lot for your review and comments.

I renamed the `test` directory to `tests` as a I thought this might better fit some conventions. I must have missed those paths. Thanks for pointing out - they're fixed now: https://github.com/oschwengers/referenceseeker/commit/96d282e9a4155a82fdabbd4ee58432506ebf04a7

I also added a `CONTRIBUTING.md` file (https://github.com/oschwengers/referenceseeker/commit/5c3949ed533b4dead927faf3a91d22f78b8d1204) which contains the requested information on how to execute the `PyTest` tests as well as two initial issue templates for bug reports (https://github.com/oschwengers/referenceseeker/commit/18163c8d580871b200f39c98172fbd6335d3e64c) and feature requests (https://github.com/oschwengers/referenceseeker/commit/9303bc86bb7fb201c9c9d7bb4ecb81c18700fee0).

I hope, everything is addressed properly. Otherwise, please let me know.

@standage , @luizirber Again, thank you very much for your very committed reviews and extreamly constructive comments and ideas!

@will-rowe , please let me know what the next steps in the process are and what I can do.
