### 2020-02-04 e819bc2
fix typo in readme
