### 2020-02-04 #1994 @oschwengers
Hi @will-rowe, thank you for the quick reply.

I just tagged a new release [v1.6](https://github.com/oschwengers/referenceseeker/releases/tag/v1.6) and archived it at Zenodo with the following DOI: [10.5281/zenodo.3635630](https://zenodo.org/record/3635630)
