### 2020-02-04 #1994 @will-rowe
Hey @openjournals/joss-eics - we're good to go with this one.

Thank you @standage and @luizirber for your excellent reviews, and thank you @oschwengers for this great submission.
