### 2020-02-04 1f51eaa
Add package build section to pythonpublish file
