+++ .github/workflows/pythonpublish.yml
@@ -17,7 +17,10 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install setuptools wheel twine
-    - name: Build and publish to PyPI
+    - name: Build package
+      run: |
+        python setup.py sdist bdist_wheel
+    - name: Publish to PyPI
       uses: pypa/gh-action-pypi-publish@master
       with:
         user: __token__
