### 2020-02-04 #1994 @oschwengers
Hi @kthyng ! Title, author and affiliations are updated now. Unfortunately, Zenodo's affiliation form is not the easiest to fill out as there is only one short text box. Please, let me know if you need something changed.
