### 2020-02-04 #1994 @oschwengers
@standage , @luizirber : Thank you very much for your great reviews! It really helped to make this software tool better.

@will-rowe Thank you very much for accepting this and all your work as the reviewer!

@kthyng Thank you for this extremely fast publication - astonishing fast :-)  Just a little bit too fast as I just found a tiny typo in the manuscript https://github.com/oschwengers/referenceseeker/commit/47b6110be33e2d4291fbd4f008bbb65e7cc73591. Would it still be possible to fix this? If not - that would also be fine.
