### 2020-02-04 #1994 @arfon
> @arfon can you help with a typo? By the way are we using the tech/dev handle yet for things like this?

This should be fixed now although it can take a few hours for the updated PDF to show on the JOSS site due to caching we have in place.

@kthyng - the process is a manual one for now unfortunately (i.e. requires @openjournals/dev to fix this).
